# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Executive Strategy Cockpit API for Rose OS / Command Center
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Metric {
  label: string;
  value: string;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  change?: string | null;
  /**
     * up, down, or flat
     * @nullable
     */
  trend?: string | null;
  /** @nullable */
  helper?: string | null;
}

export interface ExecutiveSummary {
  companyPulse: Metric;
  collectedRetainedRevenue: Metric;
  operatingReserve: Metric;
  majorDecisionsPending: Metric;
  pulseTrend: number[];
}

export interface BriefingItem {
  id: number;
  title: string;
  /** @nullable */
  detail?: string | null;
  /** on_track, watch, action_needed, in_progress */
  status: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  sourceRecord?: string | null;
}

export interface StrategyObjective {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  owner?: string | null;
  /** @nullable */
  horizon?: string | null;
  status: string;
  progress: number;
  /** @nullable */
  pillar?: string | null;
}

export interface Decision {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** The guardrail category the decision falls under */
  category: string;
  /** needs_written_approval, budget_threshold, ip_platform_rights, financial_commitment, strategic */
  approvalType: string;
  /** @nullable */
  estimatedImpact?: string | null;
  dueDate: string;
  /** pending, approved, declined, deferred */
  status: string;
  founderApprovalRequired: boolean;
  /** @nullable */
  recommendation?: string | null;
  /** @nullable */
  sourceRecord?: string | null;
}

export interface CountByLabel {
  label: string;
  count: number;
}

export interface DecisionSummary {
  total: number;
  pendingApproval: number;
  totalImpact: string;
  byApprovalType: CountByLabel[];
}

export interface DecisionNote {
  id: number;
  decisionId: number;
  author: string;
  body: string;
  createdAt: string;
}

export interface DecisionNoteInput {
  author?: string;
  /** @minLength 1 */
  body: string;
}

export interface BankSummary {
  name: string;
  balance: string;
  note: string;
}

export interface FinancialOverview {
  metrics: Metric[];
  reserveMonths: number;
  reserveTarget: number;
  revenueTriggerProgress: number;
  bankSummaries?: BankSummary[];
}

export interface MonthlyFinancialPoint {
  month: string;
  actual: number;
  plan: number;
}

export interface FinancialCommitment {
  id: number;
  label: string;
  amount: string;
  dueDate: string;
  status: string;
  approvalRequired?: boolean;
}

export interface RiskItem {
  id: number;
  name: string;
  category: string;
  /** low, medium, high */
  severity: string;
  status: string;
  openFindings?: number;
  /** @nullable */
  summary?: string | null;
}

export interface RoadmapPhase {
  id: number;
  phase: string;
  title: string;
  /** @nullable */
  description?: string | null;
  status: string;
  progress: number;
  /** @nullable */
  vendor?: string | null;
  /** @nullable */
  timeline?: string | null;
}

export interface Partnership {
  id: number;
  name: string;
  type: string;
  stage: string;
  /** @nullable */
  value?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface Milestone {
  id: number;
  title: string;
  date: string;
  /** completed, in_progress, upcoming */
  status: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface MonthlyReview {
  id: number;
  period: string;
  status: string;
  summary: string;
  highlights?: string[];
  watchItems?: string[];
  metrics?: Metric[];
}

export interface SourceRecord {
  id: number;
  title: string;
  type: string;
  source: string;
  status: string;
  clpCleared: boolean;
  lastUpdated: string;
}

export interface Guardrail {
  id: number;
  title: string;
  description: string;
  /** enforced, advisory */
  status: string;
  category: string;
}

export type GetDecisionsParams = {
status?: GetDecisionsStatus;
};

export type GetDecisionsStatus = typeof GetDecisionsStatus[keyof typeof GetDecisionsStatus];


export const GetDecisionsStatus = {
  pending: 'pending',
  approved: 'approved',
  declined: 'declined',
  deferred: 'deferred',
} as const;