# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Executive Strategy Cockpit API for Rose OS / Command Center
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /decisions:
    get:
      operationId: getDecisions
      tags: [decisions]
      summary: Major decision queue
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, declined, deferred]
      responses:
        "200":
          description: Decisions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Decision"
...
  /decisions/{id}:
    get:
      operationId: getDecision
      tags: [decisions]
      summary: Single decision detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Decision"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /decisions/{id}/notes:
    get:
      operationId: getDecisionNotes
      tags: [decisions]
      summary: Founder participation notes on a decision
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Decision notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DecisionNote"
    post:
      operationId: createDecisionNote
      tags: [decisions]
      summary: Add a founder participation note (recommend-only, not an approval)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionNoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionNote"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Decision:
      type: object
      required: [id, title, category, status, approvalType, dueDate, founderApprovalRequired]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          description: The guardrail category the decision falls under
        approvalType:
          type: string
          description: needs_written_approval, budget_threshold, ip_platform_rights, financial_commitment, strategic
        estimatedImpact:
          type: ["string", "null"]
        dueDate:
          type: string
        status:
          type: string
          description: pending, approved, declined, deferred
        founderApprovalRequired:
          type: boolean
        recommendation:
          type: ["string", "null"]
        sourceRecord:
          type: ["string", "null"]
...
    DecisionNote:
      type: object
      required: [id, decisionId, author, body, createdAt]
      properties:
        id:
          type: integer
        decisionId:
          type: integer
        author:
          type: string
        body:
          type: string
        createdAt:
          type: string
    DecisionNoteInput:
      type: object
      required: [body]
      properties:
        author:
          type: string
        body:
          type: string
          minLength: 1
...