# lib/api-zod/src/generated/types/decision.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Executive Strategy Cockpit API for Rose OS / Command Center
 * OpenAPI spec version: 0.1.0
 */

export interface Decision {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** The guardrail category the decision falls under */
  category: string;
  /** needs_written_approval, budget_threshold, ip_platform_rights, financial_commitment, strategic */
  approvalType: string;
  /** @nullable */
  estimatedImpact?: string | null;
  dueDate: string;
  /** pending, approved, declined, deferred */
  status: string;
  founderApprovalRequired: boolean;
  /** @nullable */
  recommendation?: string | null;
  /** @nullable */
  sourceRecord?: string | null;
}