# lib/api-zod/src/generated/types/error.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Executive Strategy Cockpit API for Rose OS / Command Center
 * OpenAPI spec version: 0.1.0
 */

export interface Error {
  error: string;
}